vae_opt -> vae
