+++ models/tt_dit/layers/linear.py
@@ -69,7 +69,7 @@ class Linear(Module):
         M, K, N = x.padded_shape[-2], x.padded_shape[-1], self.weight.data.padded_shape[-1]
         core_grid = get_matmul_core_grid(self.mesh_device)
 
-        if M <= 64:  # Branch B: 1D mcast_in0 matmul for small-M shapes
+        if M <= 64:  # TEMPORARY for FLUX2: Branch B: 1D mcast_in0 matmul for small-M shapes
             program_config = get_1d_matmul_config(M, K, N, core_grid)
             output = ttnn.linear(
                 x,
@@ -303,7 +303,7 @@ class ColParallelLinear(Module):
                 )
                 return [_apply_activation_fn(o, self.activation_fn) for o in outputs]
 
-            if M <= 128:  # Branch B: 1D mcast_in0 matmul for small-M shapes
+            if M <= 128:  # TEMPORARY for FLUX2: Branch B: 1D mcast_in0 matmul for small-M shapes
                 program_config = get_1d_matmul_config(M, K, N, core_grid)
                 output = ttnn.linear(
                     x,
